Separate transtiinos state to UploadMutator component
