+++ src/components/FileUpload/UploadMutator.js
@@ -1,10 +1,12 @@
 import React from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 import { useTransitions } from "redux-transitions";
-import { UPLOAD_CHUNK, UPLOAD_SUCCESS, UPLOAD_FAILURE } from "../redux/actions";
-import { uploadFile } from "../redux/thunks/uploadFile";
-
-const uploadedFilesSelector = (store) => store.upload;
+import {
+  UPLOAD_CHUNK,
+  UPLOAD_SUCCESS,
+  UPLOAD_FAILURE,
+} from "../../redux/actions";
+import { uploadFile } from "../../redux/thunks/uploadFile";
 
 const PENDING_STATE = "pending";
 const SUCCESS_STATE = "success";
@@ -22,17 +24,19 @@ const uploadReducer = (state, { error, percentage }) => ({
   uploadPercentage: percentage || 0,
 });
 
-export const FileUpload = () => {
+// this component only relies on Transition state
+// @transitionState
+// @dispatcher
+export const UploadMutator = () => {
   const dispatch = useDispatch();
 
-  const { uploadedFiles, totalSize } = useSelector(uploadedFilesSelector);
   const { isUploading, uploadPercentage, uploadError } = useTransitions(
     uploadStates,
     uploadReducer
   );
 
   return (
-    <div className="data-fetcher">
+    <>
       <button disabled={isUploading} onClick={() => dispatch(uploadFile())}>
         {isUploading ? "uploading..." : uploadError ? "retry" : "upload"}
       </button>
@@ -46,14 +50,6 @@ export const FileUpload = () => {
           Error while uploading: {uploadError}
         </span>
       )}
-      <div className="uploaded-files">
-        {uploadedFiles.map(({ fileName, fileSize }) => (
-          <div key={fileName}>
-            {fileName} [{fileSize.toFixed(2)} Mb]
-          </div>
-        ))}
-      </div>
-      Total Size: {totalSize.toFixed(2)} Mb
-    </div>
+    </>
   );
 };
