+++ src/components/FileUpload/index.js
@@ -1,0 +1,3 @@
+import { FileUpload } from "./FileUpload";
+
+export default FileUpload;
